util: refactor LazyLoadBalancer into util to use it in autosharding
