+++ util/build.gradle
@@ -56,6 +56,7 @@ animalsniffer {
 }
 
 tasks.named("javadoc").configure {
+    exclude 'io/grpc/util/LazyLoadBalancer*'
     exclude 'io/grpc/util/MultiChildLoadBalancer.java'
     exclude 'io/grpc/util/OutlierDetectionLoadBalancer*'
     exclude 'io/grpc/util/RandomSubsettingLoadBalancer*'
