+++ xds/src/main/java/io/grpc/xds/RingHashLoadBalancer.java
@@ -39,6 +39,7 @@ import io.grpc.LoadBalancer;
 import io.grpc.Metadata;
 import io.grpc.Status;
 import io.grpc.SynchronizationContext;
+import io.grpc.util.LazyLoadBalancer;
 import io.grpc.util.MultiChildLoadBalancer;
 import io.grpc.xds.ThreadSafeRandom.ThreadSafeRandomImpl;
 import io.grpc.xds.client.XdsLogger;
